+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for 4:2:2 and 4:4:4 pixel formats (8 and 10 bits) in VideoToolbox
   decoder
 
+### Deprecated
+- `skip` and `trim_duration` are deprecated over `start_time` and `end_time`
+  parameters
+
 ## [9.11.1] - 2022-02-10
 ### Fixed
 - Fix build with FFmpeg >= 5.0